Improvements to diag info (#666)
* Improve diag info

* Rename message outgress to egress

* Remove processing/sending stats

* Add percent to sent/dropped and calculated OPC values to queues

* Add thousand separator

* Fix dropped percent

* Show estimated messages only when > 0

* Show % consistently with space separator

* Show OPC values estimate for egress queue

* Fix property name MaxEgressMessageQueue

* Cosmetic changes to output

* Replace Collection with List

* Set message counter threshold to ulong

* Remove percentage from hub queue

* Add method to calc ValueChanges percentage

* Revert "Replace Collection with List"

This reverts commit 50b3e3d68c3f9b54142ec7a07c02ebcc5836214e.

* Set max efress msg queue default to 4096

* Rename PCS_DEFAULT_PUBLISH_MAX_OUTGRESS_MESSAGES to PCS_DEFAULT_PUBLISH_MAX_EGRESS_MESSAGE_QUEUE

* Add divide by zero check

* MaxEgressMessageQueue default to 4096 bytes

+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Edge.Publisher/src/Engine/IoTHubMessageSink.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Edge.Publisher.Engine {
     public class IoTHubMessageSink : IMessageSink, IDisposable {
 
         /// <inheritdoc/>
-        public long SentMessagesCount { get; private set; }
+        public ulong SentMessagesCount { get; private set; }
 
         /// <summary>
         /// Create IoT hub message sink
@@ -81,9 +81,8 @@ namespace Microsoft.Azure.IIoT.OpcUa.Edge.Publisher.Engine {
                     sw.Stop();
                     _logger.Verbose("Sent {count} messages in {time} to IoTHub.", messagesCount, sw.Elapsed);
                 }
-                SentMessagesCount += messagesCount;
-                kMessagesSent.WithLabels(IotHubMessageSinkGuid,
-                    IotHubMessageSinkStartTime).Set(SentMessagesCount);
+                SentMessagesCount += (ulong)messagesCount;
+                kMessagesSent.WithLabels(IotHubMessageSinkGuid, IotHubMessageSinkStartTime).Set(SentMessagesCount);
             }
             catch (Exception ex) {
                 _logger.Error(ex, "Error while sending messages to IoT Hub."); // we do not set the block into a faulted state.
@@ -123,7 +122,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Edge.Publisher.Engine {
             return msg;
         }
 
-        private const long kMessageCounterResetThreshold = long.MaxValue - 10000;
+        private const ulong kMessageCounterResetThreshold = ulong.MaxValue - 10000;
         private readonly ILogger _logger;
         private readonly IClientAccessor _clientAccessor;
         private readonly string IotHubMessageSinkGuid = Guid.NewGuid().ToString();
